requirements.txt
fastapi
uvicorn
selenium
google-auth
google-auth-oauthlib
google-api-python-client
python-multipart
